+++ packages/docs/.vitepress/config/shared.ts
@@ -8,16 +8,26 @@ if (process.env.NETLIFY) {
   console.log('Netlify build', process.env.CONTEXT)
 }
 
-const productionHead: HeadConfig[] = [
-  [
-    'script',
-    {
-      src: 'https://unpkg.com/thesemetrics@latest',
-      async: '',
-      type: 'text/javascript',
-    },
-  ],
-]
+const productionHead: HeadConfig[] = []
+
+const rControl = /[\u0000-\u001f]/g
+const rSpecial = /[\s~`!@#$%^&*()\-_+=[\]{}|\\;:"'“”‘’<>,.?/]+/g
+const rCombining = /[\u0300-\u036F]/g
+
+/**
+ * Default slugification function
+ */
+export const slugify = (str: string): string =>
+  str
+    .normalize('NFKD')
+    // Remove accents
+    .replace(rCombining, '')
+    // Remove control characters
+    .replace(rControl, '')
+    // Replace special characters
+    .replace(rSpecial, '-')
+    // ensure it doesn't start with a number
+    .replace(/^(\d)/, '_$1')
 
 export const sharedConfig = defineConfig({
   title: 'Pinia',
@@ -32,6 +42,10 @@ export const sharedConfig = defineConfig({
     attrs: {
       leftDelimiter: '%{',
       rightDelimiter: '}%',
+    },
+
+    anchor: {
+      slugify,
     },
   },
 
@@ -67,16 +81,26 @@ export const sharedConfig = defineConfig({
       },
     ],
 
-    // TODO: add this back when fixed
-    // [
-    //   'script',
-    //   {
-    //     src: 'https://vueschool.io/banners/main.js',
-    //     // @ts-expect-error: vitepress bug
-    //     async: true,
-    //     type: 'text/javascript',
-    //   },
-    // ],
+    [
+      'script',
+      {
+        src: 'https://cdn.usefathom.com/script.js',
+        'data-site': 'KFPPRRIS',
+        'data-spa': 'auto',
+        defer: '',
+      },
+    ],
+
+    // Vue School Top banner
+    [
+      'script',
+      {
+        src: 'https://vueschool.io/banner.js?affiliate=vuerouter&type=top',
+        // @ts-expect-error: vitepress bug
+        async: true,
+        type: 'text/javascript',
+      },
+    ],
 
     ...(isProduction ? productionHead : []),
   ],
@@ -86,7 +110,7 @@ export const sharedConfig = defineConfig({
     outline: [2, 3],
 
     socialLinks: [
-      { icon: 'twitter', link: 'https://twitter.com/posva' },
+      { icon: 'x', link: 'https://twitter.com/posva' },
       {
         icon: 'github',
         link: 'https://github.com/vuejs/pinia',