+++ packages/pinia/src/rootStore.ts
@@ -1,8 +1,8 @@
 import {
   App,
   EffectScope,
-  getCurrentInstance,
   inject,
+  hasInjectionContext,
   InjectionKey,
   Ref,
 } from 'vue-demi'
@@ -32,15 +32,21 @@ export let activePinia: Pinia | undefined
  *
  * @param pinia - Pinia instance
  */
-export const setActivePinia = (pinia: Pinia | undefined) =>
-  (activePinia = pinia)
+// @ts-expect-error: cannot constrain the type of the return
+export const setActivePinia: _SetActivePinia = (pinia) => (activePinia = pinia)
+
+interface _SetActivePinia {
+  (pinia: Pinia): Pinia
+  (pinia: undefined): undefined
+  (pinia: Pinia | undefined): Pinia | undefined
+}
 
 /**
  * 获取当前活跃的pinia实例
  * Get the currently active pinia if there is any.
  */
 export const getActivePinia = () =>
-  (getCurrentInstance() && inject(piniaSymbol)) || activePinia
+  (hasInjectionContext() && inject(piniaSymbol)) || activePinia
 
 /**
  * 每个应用程序必须拥有属性自己的pinia来创建仓库