+++ rollup.config.mjs
@@ -1,20 +1,28 @@
 // @ts-check
-import path from 'path'
+import { dirname, resolve } from 'node:path'
+import { fileURLToPath } from 'node:url'
+import { readFileSync } from 'node:fs'
 import ts from 'rollup-plugin-typescript2'
 import replace from '@rollup/plugin-replace'
-import resolve from '@rollup/plugin-node-resolve'
+import nodeResolve from '@rollup/plugin-node-resolve'
 import commonjs from '@rollup/plugin-commonjs'
 import pascalcase from 'pascalcase'
 import terser from '@rollup/plugin-terser'
+import chalk from 'chalk'
+
+const __filename = fileURLToPath(import.meta.url)
+const __dirname = dirname(__filename)
 
 if (!process.env.TARGET) {
   throw new Error('TARGET package must be specified via --environment flag.')
 }
 
-const packagesDir = path.resolve(__dirname, 'packages')
-const packageDir = path.resolve(packagesDir, process.env.TARGET)
+const packagesDir = resolve(__dirname, 'packages')
+const packageDir = resolve(packagesDir, process.env.TARGET)
 
-const pkg = require(path.resolve(packageDir, `package.json`))
+const pkg = JSON.parse(
+  readFileSync(resolve(packageDir, `package.json`), 'utf-8')
+)
 const name = pkg.name
 
 function getAuthors(pkg) {
@@ -31,10 +39,10 @@ function getAuthors(pkg) {
 }
 
 const banner = `/*!
-  * ${pkg.name} v${pkg.version}
-  * (c) ${new Date().getFullYear()} ${getAuthors(pkg)}
-  * @license MIT
-  */`
+ * ${pkg.name} v${pkg.version}
+ * (c) ${new Date().getFullYear()} ${getAuthors(pkg)}
+ * @license MIT
+ */`
 
 // ensure TS checks only once for each build
 let hasTSChecked = false
@@ -80,7 +88,7 @@ export default packageConfigs
 
 function createConfig(buildName, output, plugins = []) {
   if (!output) {
-    console.log(require('chalk').yellow(`invalid format: "${buildName}"`))
+    console.log(chalk.yellow(`invalid format: "${buildName}"`))
     process.exit(1)
   }
 
@@ -105,15 +113,15 @@ function createConfig(buildName, output, plugins = []) {
 
   const tsPlugin = ts({
     check: !hasTSChecked,
-    tsconfig: path.resolve(__dirname, './tsconfig.json'),
-    cacheRoot: path.resolve(__dirname, './node_modules/.rts2_cache'),
+    tsconfig: resolve(__dirname, './tsconfig.json'),
+    cacheRoot: resolve(__dirname, './node_modules/.rts2_cache'),
     tsconfigOverride: {
       compilerOptions: {
         sourceMap: output.sourcemap,
         declaration: shouldEmitDeclarations,
         declarationMap: shouldEmitDeclarations,
       },
-      exclude: ['packages/*/__tests__', 'packages/*/test-dts'],
+      exclude: ['*.spec.ts', 'packages/*/test-dts', 'packages/*/testing'],
     },
   })
   // we only need to check TS and generate declarations once for each build.
@@ -130,7 +138,7 @@ function createConfig(buildName, output, plugins = []) {
     external.push('@vue/devtools-api')
   }
 
-  const nodePlugins = [resolve(), commonjs()]
+  const nodePlugins = [nodeResolve(), commonjs()]
 
   return {
     input: `src/index.ts`,
@@ -166,23 +174,29 @@ function createReplacePlugin(
   isGlobalBuild,
   isNodeBuild
 ) {
+  const __DEV__ =
+    (isBundlerESMBuild && !isRawESMBuild) || (isNodeBuild && !isProduction)
+      ? // preserve to be handled by bundlers
+        `(process.env.NODE_ENV !== 'production')`
+      : // hard coded dev/prod builds
+        JSON.stringify(!isProduction)
+  const __FEATURE_PROD_DEVTOOLS__ = isBundlerESMBuild
+    ? `(typeof __VUE_PROD_DEVTOOLS__ !== 'undefined' && __VUE_PROD_DEVTOOLS__)`
+    : 'false'
+
+  const __TEST__ =
+    (isBundlerESMBuild && !isRawESMBuild) || isNodeBuild
+      ? `(process.env.NODE_ENV === 'test')`
+      : 'false'
+
   const replacements = {
     __COMMIT__: `"${process.env.COMMIT}"`,
     __VERSION__: `"${pkg.version}"`,
-    __DEV__:
-      (isBundlerESMBuild && !isRawESMBuild) || (isNodeBuild && !isProduction)
-        ? // preserve to be handled by bundlers
-          `(process.env.NODE_ENV !== 'production')`
-        : // hard coded dev/prod builds
-          JSON.stringify(!isProduction),
+    __USE_DEVTOOLS__: `((${__DEV__} || ${__FEATURE_PROD_DEVTOOLS__}) && !${__TEST__})`,
+    __DEV__,
     // this is only used during tests
-    __TEST__:
-      (isBundlerESMBuild && !isRawESMBuild) || isNodeBuild
-        ? `(process.env.NODE_ENV === 'test')`
-        : 'false',
-    __FEATURE_PROD_DEVTOOLS__: isBundlerESMBuild
-      ? `(typeof __VUE_PROD_DEVTOOLS__ !== 'undefined' && __VUE_PROD_DEVTOOLS__)`
-      : 'false',
+    __TEST__,
+    __FEATURE_PROD_DEVTOOLS__,
     // If the build is expected to run directly in the browser (global / esm builds)
     __BROWSER__: JSON.stringify(isRawESMBuild),
     // is targeting bundlers?